+++ driver.go
@@ -497,9 +497,9 @@ func (d *Driver) StartTask(cfg *drivers.TaskConfig) (*drivers.TaskHandle, *drive
 	createOpts.ContainerBasicConfig.Labels = driverConfig.Labels
 
 	// Logging
-	if driverConfig.Logging.Driver != "" {
+	if !driverConfig.Logging.Empty() {
 		switch driverConfig.Logging.Driver {
-		case LOG_DRIVER_NOMAD:
+		case "", LOG_DRIVER_NOMAD:
 			if !d.config.DisableLogCollection {
 				createOpts.LogConfiguration.Driver = "k8s-file"
 				createOpts.ContainerBasicConfig.LogConfiguration.Path = cfg.StdoutPath
@@ -513,9 +513,7 @@ func (d *Driver) StartTask(cfg *drivers.TaskConfig) (*drivers.TaskHandle, *drive
 	} else {
 		d.logger.Trace("no podman log driver provided, defaulting to plugin config")
 		switch d.config.Logging.Driver {
-		case "": // The default is nomad
-			fallthrough
-		case LOG_DRIVER_NOMAD:
+		case "", LOG_DRIVER_NOMAD:
 			if !d.config.DisableLogCollection {
 				createOpts.LogConfiguration.Driver = "k8s-file"
 				createOpts.ContainerBasicConfig.LogConfiguration.Path = cfg.StdoutPath